+++ types/permissionTypes.ts
@@ -1,0 +1,5 @@
+export const enum Permission {
+    Host = 1 << 0,
+    PlaybackControl = 1 << 1,
+    QueueControl = 1 << 2,
+  }
